+++ CHANGES.md
@@ -5,7 +5,7 @@ v1.0.4 (TBD)
 ===================
 Non-breaking changes:
 ---------------------
-*
+* New `--txt-extension` argument to add a `.txt` extension to the output files' names.
 
 v1.0.3 (2020-09-30)
 ===================